feat(shows): add static assets for fashion shows
This patch adds the static assets for fashion shows into this separate
`static` repository that will be deployed to Vercel independent from the
original `site` repository (where the new fashion show stuff is being
worked on [[1]]). This is a temporary solution that ultimately should be
replaced by some cloud storage solution (e.g. S3, Vercel Blob [[2]], or
Supabase Storage [[3]]). For now, this is the free way to do it.

[1]: https://github.com/nicholaschiang/site/pull/23
[2]: https://vercel.com/docs/storage/vercel-blob
[3]: https://supabase.com/docs/guides/storage